Updated following packages to latest git version:
xserver          commit f59871587ea678d4c498874895f4b97687764ee1
libxcb           commit 3c946010c8521497b0fba2c8bc9bde184622345a
libxcb/xcb-proto commit 1388374c7149114888a6a5cd6e9bf6ad4b42adf8
xkeyboard-config commit 3998ddde5bda96538fadd7de42a63f0b7c43d85d
libX11           commit c745719e23af44a4b40ab4508447637b35d91a1e
libXdmcp         commit fa41f9bfd7a0e57ba3e0f40d9a96e1dba1deeb4a
libXext          commit 6cb21433d745abea3a161a6fc8e141f7e08b2c27
libfontenc       commit 2408433404732b146bf41a2451d23ffadad5a816
libXinerama      commit 51c28095951676a5896437c4c3aa40fb1972bad2
libXau           commit f4fc44202b554e68093828b556f8c7fd9644ddb0
xkbcomp          commit ff32f542c94f1b30ba2ba58193adb3774a5c50e6
pixman           commit 7883ab8d637cbdf3b12bb17525616a5372de3d02
mkfontscale      commit b7ebb9b042974874782f4e662bfc442bc96c5647
xwininfo         commit ab427b9bfd495a7e9a5dc690f5abe5a2859b8de3
libXft           commit 4e4210c86019b42c3819cc18b59f86752fcd8b13
libXmu           commit 017d1deaf923887df74b068c33d080008b0c25cf
libxtrans        commit 806f04c6e4529358f160e53135baf105e4ecf3b8
fontconfig       commit 72b9a48f57de6204d99ce1c217b5609ee92ece9b
mesa             commit 55c262898ae7188311c89a60e4ec0fbb67b7a95b
putty            commit b846178443cf1a5dc7c5ea2079fd34fd465af497
pthreads         commit f12b445b336ee0117b43fca1d4b9f22c9af82c36
EGL-Registry     commit a03692eea13514d9aef01822b2bc6575fcabfac2
xorgproto        commit 1c8128d72df22843a2022576850bc5ab5e3a46ea

+++ libXmu/src/StdCmap.c
@@ -190,11 +190,11 @@ valid_args(XVisualInfo *vinfo, unsigned long red_max, unsigned long green_max,
 	    return 0;
     } else if (property == XA_RGB_GRAY_MAP) {
 	ncolors = red_max + green_max + blue_max + 1;
-	if (ncolors > vinfo->colormap_size)
+	if (ncolors > (unsigned long) vinfo->colormap_size)
 	    return 0;
     } else {
 	ncolors = (red_max + 1) * (green_max + 1) * (blue_max + 1);
-	if (ncolors > vinfo->colormap_size)
+	if (ncolors > (unsigned long) vinfo->colormap_size)
 	    return 0;
     }
